Adding "the right solution"
This is the solution from the class slides.  This solution creates a data class called unit which has a number of units. The unit data type then derives with the Ord and Eq type classes and then defines how each unit should be shown.

The factor function defines the conversion constants. Note that it defines the last three units as a reciprocal of the first three units. The inverse class then maps a  unit to its reciprocal unit. Finally, the convert function returns a tuple  with the correct conversion.